Add NoInit inventory getters
The implementation for GetInventoryChanges in CLIB does not match that of the game; the game's does not attempt to initialize the inventory if it hasn't been already.

Initializing the inventory can have side-effects, like initializing the leveled items. This is undesirable in many circumstances, so we add NoInit versions of the inventory getters to prevent this.